finished TODO corrections

+++ src/main/java/com/mc/HouseManagement/entity/Owner.java
@@ -12,11 +12,11 @@ public class Owner extends Person{
     public Owner() {
     }
 
-    public Owner(String firstName, String lastName, String email, Integer phone, List<Apartment> apartments) {
+    public Owner(String firstName, String lastName, String email, Long phone, List<Apartment> apartments) {
         super(firstName, lastName, email, phone, apartments);
     }
 
-    public static Owner createOwner(String firstName, String lastName, String email, Integer phone, List<Apartment> apartments){
+    public static Owner createOwner(String firstName, String lastName, String email, Long phone, List<Apartment> apartments){
         return new Owner(firstName, lastName, email, phone, apartments);
     }
 }
