+++ src/main/java/com/mc/HouseManagement/entity/SoldMovedOut.java
@@ -11,11 +11,11 @@ public class SoldMovedOut extends Person{
 
     public SoldMovedOut(){}
 
-    public SoldMovedOut(String firstName, String lastName, String email, Integer phone, List<Apartment> apartments){
+    public SoldMovedOut(String firstName, String lastName, String email, Long phone, List<Apartment> apartments){
         super(firstName,lastName,email,phone,apartments);
     }
 
-    public static SoldMovedOut createSoldMovedOut(String firstName, String lastName, String email, Integer phone, List<Apartment> apartments){
+    public static SoldMovedOut createSoldMovedOut(String firstName, String lastName, String email, Long phone, List<Apartment> apartments){
         return new SoldMovedOut(firstName,lastName,email,phone,apartments);
     }
 }
