+++ src/main/java/com/mc/HouseManagement/entity/User.java
@@ -11,11 +11,11 @@ public class User extends Person{
     public User() {
     }
 
-    public User(String firstName, String lastName, String email, Integer phone, List<Apartment> apartments) {
+    public User(String firstName, String lastName, String email, Long phone, List<Apartment> apartments) {
         super(firstName, lastName, email, phone, apartments);
     }
 
-    public static User createUser(String firstName, String lastName, String email, Integer phone, List<Apartment> apartments){
+    public static User createUser(String firstName, String lastName, String email, Long phone, List<Apartment> apartments){
         return new User(firstName, lastName, email, phone, apartments);
     }
 }
